Improve Windows build and VSCode integration
Use the built-in generator "$<CONFIG>" instead of defining a separate
"PROJECT_CONFIG" variable; this improves compatibility with "generator"
type build systems like Visual Studio. As well, update install
directories for the third_party includes so that they're included in the
build/ folder and not just installed to the root of the repository.

Finally, remove some stale comments and remove the default LIBCMTD
that's used when the Debug config is used on Windows.

I need to test this whole change on FreeBSD before I merge this to the
main branch.

Signed-off-by: Eric Joyner <erj@erj.cc>