Fix the stride used when writing out decoded images
The stride from the decoded image returned by libheif does not
necessarily match the width of the final output image; so the
final output image's width should not be used as the stride when
reading from the decoded image's memory buffer.

Fixes the decoding for the resized 1620x1080 image that I've been using
to test.

Signed-off-by: Eric Joyner <erj@erj.cc>